[1.0.1] Enforcing stronger incoming value checking.
When building or requesting routes, when initializing the respective modules, checks will be preformed on the passed routes JSON file or endpoint directory for existence and format before proceeding.
